Fix for tabular platform tests on Windows runner (#3493)

+++ tabular/tests/unittests/test_tabular.py
@@ -37,6 +37,7 @@ from autogluon.tabular import TabularDataset, TabularPredictor
 
 PARALLEL_LOCAL_BAGGING = "parallel_local"
 SEQUENTIAL_LOCAL_BAGGING = "sequential_local"
+on_windows = os.name == "nt"
 
 
 def test_tabular():
@@ -137,7 +138,8 @@ def test_advanced_functionality():
     savedir_predictor_original = savedir + "predictor/"
     predictor: TabularPredictor = TabularPredictor(label=label, path=savedir_predictor_original).fit(train_data)
     leaderboard = predictor.leaderboard(data=test_data)
-    predictor.plot_ensemble_model()
+    if not on_windows:
+        predictor.plot_ensemble_model()
     extra_metrics = ["accuracy", "roc_auc", "log_loss"]
     test_data_no_label = test_data.drop(columns=[label])
     with pytest.raises(ValueError):
@@ -247,7 +249,8 @@ def test_advanced_functionality():
 
     assert predictor.get_model_full_dict() == dict()
     predictor.refit_full()
-    predictor.plot_ensemble_model()
+    if not on_windows:
+        predictor.plot_ensemble_model()
     assert len(predictor.get_model_full_dict()) == num_models
     assert len(predictor.get_model_names()) == num_models * 2
     for model in predictor.get_model_names():
@@ -1143,7 +1146,8 @@ def test_tabular_log_to_file():
     predictor = TabularPredictor(label="class", log_to_file=True, log_file_path=log_file).fit(train_data=train_data, hyperparameters={"DUMMY": {}})
     log = TabularPredictor.load_log(log_file_path=log_file)
     assert "TabularPredictor saved." in log[-1]
-    os.remove(log_file)
+    if not on_windows:
+        os.remove(log_file)
 
     with pytest.raises(AssertionError):
         TabularPredictor.load_log()
